// src/app/models/APIDetails.ts
import { APIDefinition } from "./APIDefinition";

export class APIDetails {

	id: string;
	name: string;
	description: string;
	context: string;
	version: string;
	provider: string;
	status: string;
	thumbnailUri: string;
	accessControl: string;
	visibility: string;
	tags: string;
	apiDefinition: string;
	wsdlUri: string;
	responseCaching: string;
	cacheTimeout: string;
	destinationStatsEnabled: string;
	isDefaultVersion: string;
	type: string;

}